+++ app/src/test/java/com/github/komidawi/pizzacostcalculator/screen/add/AddPizzaFragmentViewModelTest.kt
@@ -121,7 +121,7 @@ class AddPizzaFragmentViewModelTest {
         val ratio = viewModel.calculateRatio()
 
         // then
-        assertEquals(testRatio, ratio.toFloat(), delta)
+        assertEquals(testRatio, ratio.toDouble(), delta)
     }
 
     @Test
